Fix import errors and update Desk Structure
This pull request fixes some import errors in the code by removing unused imports and adding an import for a missing function getFilteredDocumentTypeListItems. Additionally, I updated the Desk Structure code to remove the client that wasn't used and I made explicit what the i18nConfig is. This update improves the overall clarity and ease of the docs. I tested the changes locally and verified that everything is working correctly.